+++ R/assess.R
@@ -29,13 +29,25 @@
 #'
 #' @details Until relatively recently the standard way of testing for a difference between a case and controls was to convert the case’s score to a z score using the control sample mean and standard deviation (SD). If z was less than -1.645 (i.e., below 95% of the controls) then it was concluded that the case was significantly lower than controls. However, this method has serious disadvantages (Crawford and Garthwaite, 2012).
 #'
-#' Crawford et al. (2012) remind that empirically based decision algorithms are almost always vastly superior to clinically based decision making while being more reliable, accurate, and cost-effective (Dawes, Faust, & Meehl, 1989; Grove & Lloyd, 2006; Salzinger, 2005).
-#'
 #' @importFrom stats ecdf
 #' @import ggplot2
 #' @import dplyr
 #' @export
-assess <- function(patient, mean = 0, sd = 1, n = NULL, controls = NULL, CI=95, treshold=0.05, iter=10000, color_controls="#2196F3", color_CI="#E91E63", color_score="black", color_size=2, alpha_controls=1, alpha_CI=0.8, verbose=TRUE) {
+assess <- function(patient,
+                   mean = 0,
+                   sd = 1,
+                   n = NULL,
+                   controls = NULL,
+                   CI=95,
+                   treshold=0.05,
+                   iter=10000,
+                   color_controls="#2196F3",
+                   color_CI="#E91E63",
+                   color_score="black",
+                   color_size=2,
+                   alpha_controls=1,
+                   alpha_CI=0.8,
+                   verbose=TRUE) {
   if (is.null(controls)) {
     if (is.null(n)) {
       if (verbose == TRUE) {
@@ -54,12 +66,42 @@ assess <- function(patient, mean = 0, sd = 1, n = NULL, controls = NULL, CI=95, 
       warning("Multiple scores were provided. Returning a list of results.")
     }
     results <- list()
-    for (i in 1:length(patient)) {
-      results[[i]] <- crawford.test(patient[i], controls, mean, sd, n, CI, treshold, iter, color_controls, color_CI, color_score, color_size, alpha_controls, alpha_CI)
+    for (i in seq_len(length(patient))) {
+      results[[i]] <- crawford.test(
+        patient[i],
+        controls,
+        mean,
+        sd,
+        n,
+        CI,
+        treshold,
+        iter,
+        color_controls,
+        color_CI,
+        color_score,
+        color_size,
+        alpha_controls,
+        alpha_CI
+      )
       return(results)
     }
   } else {
-    result <- crawford.test(patient, controls, mean, sd, n, CI, treshold, iter, color_controls, color_CI, color_score, color_size, alpha_controls, alpha_CI)
+    result <- crawford.test(
+      patient,
+      controls,
+      mean,
+      sd,
+      n,
+      CI,
+      treshold,
+      iter,
+      color_controls,
+      color_CI,
+      color_score,
+      color_size,
+      alpha_controls,
+      alpha_CI
+    )
     return(result)
   }
 }